Add comment into each classes

+++ src/util/Gacha.java
@@ -1,8 +1,12 @@
 package util;
 
 public class Gacha {
+    /**
+     * This function is to get the random number between 0 and 1
+     * @param min 0
+     * @param max 1
+     */
     public static Double getRandomNumber(double min, double max) {
-        return (double) ((Math.random() * (max - min)) + min);
-
+        return (Math.random() * (max - min)) + min;
     }
 }
